+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2.2.0-rc2
 
 - Remote: use HTTP status code for what is considered valid or not (#956)
+- Remote: allow RESTful urls where "{value}" is replaced by the value to validate
 - Remote: add field:ajaxoptions to allow customizing of the ajax parameters (#894)
 - pattern validator is now anchored, unless it looks like /pattern/flag (#861)